feat(warehouse): add warehouse list api

+++ app/common/models/warehouse.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, SmallInteger, DECIMAL, String, Enum, text
+from sqlalchemy import Column, Integer, SmallInteger, String, Enum, text, Float
 from app.common.types.warehouse import WarehouseAction
 from database.base_model import BaseModel
 
@@ -24,7 +24,7 @@ class WarehouseIn(BaseModel):
 
     __tablename__ = "warehouse_in"
     goods_id = Column(Integer, nullable=False, comment="商品ID")
-    price = Column(DECIMAL, nullable=False, comment="商品单价")
+    price = Column(Float, nullable=False, comment="商品单价")
     in_num = Column(SmallInteger, nullable=False, comment="入库数量")
     in_type = Column(Enum(WarehouseAction))
     in_code = Column(String(30), nullable=False, comment="入库单号")
@@ -51,5 +51,5 @@ class WarehouseHistory(BaseModel):
     goods_id = Column(Integer, nullable=False, comment="商品ID")
     num = Column(SmallInteger, nullable=False, comment="出入库数量")
     type = Column(Enum(WarehouseAction))
-    price = Column(DECIMAL, nullable=False, comment="商品单价")
+    price = Column(Float, nullable=False, comment="商品单价")
     code = Column(String(30), nullable=False, comment="出入库单号")
